[REVERTME] horrible_hacks: copy a bunch of hss clock config functions
This just forklifts a bunch of definitions etc. from hss-system-services.

This stuff needs to be removed and the corresponding initializations moved to proper functions.
Now it is all put into a single place in "clockinit".

hss-system-services is MIT -licensed, so it is ok to use it as a reference like this

The register definitions etc. need to be written properly in nuttx style

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>