Prepare riscv / rv64 for CONFIG_BUILD_KERNEL
Setup the correct amount of call gates and the parameters passed there
for kernel build. As far as I know, the same call gates are needed
for protected / kernel builds.

When enabling / testing the kernel build, revisit the riscv_swint
file and ensure the ABI is met. Testing the call gates is not possible
right now.