risc-v/mpfs: ihc: cleanup DEBUGASSERTs and reboot
Replace DEBUGASSERTs with sanity checks. DEBUGASSERT()s are
not necessarily enabled at all, thus risking the functionality
especially in that case. Remove PANICs as well.

Don't enable the ihc irq too early. If enabled, and the master
is already up, the irq is being issued so that the system gets
stuck or is severely slowed down. Master may be already if this
NuttX hart is rebooted, for example.

Signed-off-by: Eero Nurkkala <eero.nurkkala@offcode.fi>