nxxx_clockconfig.c: Disable GDET0/1 to prevent system reset
When changing the system clock to 150MHz a glitch in the system power
will occur when we configure the LDO to high drive strength.

This will cause a tamper detect related system reset, which is not what
we want when we are configuring the clock. Prevent this by disabling
the glitch detectors.