sched/sched_setpriority: Change SMP call aynchronous
Using synchronous SMP call to set priority for a task on another CPU is wrong.
Setpriority is called from semaphore priority inheritance code, from
sem_wait->boost priority, and the sem_wait can't be called recursively.

There is, however, no need to wait for the SMP call to complete, so just
change it to asynchronous.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>