sched: For SMP, don't select a CPU which is running a sched-locked tcb
In nxsched_mergepending, treat a sched locked task in the way as if the task
was running on the highest possible priority.

This allows a task to be scheduled to run on another CPU, and prevents a task
being completely locked out from scheduling, if there is a lower-priority task
running scheduler locked.

Even though scheduler being locked is a transient state that might change on
the other CPU, it is ok to check it in mergepending. If the scheduler
happens to be locked at the time of checking, the task running the locked
task will re-schedule when it exits the sched lock. If the
task on the other CPU happens to lock the scheduler before the new
task is scheduled in, that CPU will just push the task back to the
ready-to-run list.

For the rare case that all the CPUs are running sched-locked tasks, allow the
nxsched_select_cpu to return an invalid CPU number, and check that in places
where CPU selection is done.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>