sched: Optimize nxsched_remove_self and sched_unlock
Both of these are hot functions used from semaphore code. They both only need
to select a new task to run on the current CPU, and don't need to do full
scheduling accross all CPUs.

So just check if there is an eglible task in ready-to-run list, and switch it
to run on current CPU, if needed.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>