[NOT UPSTREAM] sched/wqueue/kwork_queue.c: Move sem_post outside of spinlock
This issue should not be present in upstream NuttX, so this patch should be
dropped when rebasing next time.

work_queue_period_wq is called with critical_section held from addrenv_switch
(deferred switch). But typically it is called without critical_section, and
it may post the semaphore, which takes the critical section.

This may lead to deadlock on an SMP system:
1. CPU0 adds a work, locking wqueue->lock
2. CPU1 does a deferred addrenv_switch, entering critical_section
3. CPU0 does sem_post, spining at critical section entry (which is locked by CPU1)
4. CPU1 adds a work, spinning in wqueue->lock (which is locked by CPU0)

But there should be no reason to do the sem_post while holding the wqueue->lock.
First free the spinlock and then post.

This allows the work_queue to be called from inside critical_section as well
without danger of deadlock.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>